+++ lib/get_connect/http/src/http/stub/file_decoder_stub.dart
@@ -1,0 +1,5 @@
+void writeOnFile(List<int> bytes) {}
+
+List<int> fileToBytes(dynamic data) {
+  throw UnimplementedError();
+}